all-products component rendering, changed name of main component to navbar
